pyproject.toml
[project]
name = "StochasticPaSR"
version = "0.1.0"
requires-python = "==3.10.*"

dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "matplotlib>=3.7",
    "scipy>=1.11",
    "cantera>=3.0",
    "pyyaml>=6.0",
]
